Resolve issues #4 and #6 and minor changes
Resolve issues #4 and #6 as discussed in working session. In addition, the on, once, and off methods are no longer labeled as events (they are methods for attaching event handlers, not actually events.) A clarification remark is added to the "rejected" event, as it is currently the only event associated with the end of a call where the "disconnected" event is not also triggered. It is recommended to format the event string documentation more cleanly, as currently compiling the jsdoc results in a large unformatted block of text.